Create randomizedCSVSplittah.py
this script takes in a csv file and divides or splits it into multiple files (specified by the user). What's special about this script is that it shuffles the original input rows... so your output is randomized. If you don't like this functionality and would like your output rows to retain the order of the input, just remove or comment out the line random.shuffle(lines)

Enjoy!